Initial release of adapter using node require()
