+++ Gruntfile.js
@@ -6,6 +6,17 @@
 module.exports = function(grunt) {
   grunt.initConfig({
     pkg: grunt.file.readJSON('package.json'),
+    browserify: {
+      js: {
+        src: ['src/js/adapter_core.js'],
+        dest: 'adapter.js'
+      },
+      options: {
+        browserifyOptions: {
+          standalone: 'adapter'
+        }
+      }
+    },
     githooks: {
       all: {
         'pre-commit': 'jshint jscs'
@@ -15,10 +26,10 @@ module.exports = function(grunt) {
       options: {
         jshintrc: '.jshintrc'
       },
-      files: ['adapter.js', 'test/*.js']
+      files: ['adapter_core.js', 'test/*.js']
     },
     jscs: {
-      src: ['adapter.js', 'test/*.js'],
+      src: ['adapter_core.js', 'test/*.js'],
       options: {
         config: '.jscsrc',
         'excludeFiles': [
@@ -33,8 +44,9 @@ module.exports = function(grunt) {
   grunt.loadNpmTasks('grunt-contrib-jshint');
   grunt.loadNpmTasks('grunt-jscs');
   grunt.loadNpmTasks('grunt-githooks');
+  grunt.loadNpmTasks('grunt-browserify');
   grunt.registerTask('verify-require', 'Verifies the script can be required in a node context', function () {
       require('./adapter');
   });
-  grunt.registerTask('default', ['jshint', 'jscs', 'verify-require']);
+  grunt.registerTask('default', ['jshint', 'jscs', 'browserify', 'verify-require']);
 };
