+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.2.5",
   "description": "Hide browser differences in WebRTC APIs (test package name)",
   "license": "BSD-3-Clause",
-  "main": "adapter.js",
+  "main": "adapter-core.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/webrtc/adapter.git"
@@ -18,10 +18,10 @@
     "browserify": "^10.2.1",
     "chromedriver": "^2.16.0",
     "grunt": "^0.4.5",
+    "grunt-browserify": "^4.0.1",
     "grunt-cli": ">=0.1.9",
     "grunt-contrib-jshint": "^0.11.2",
     "grunt-contrib-nodeunit": "~0.4.1",
-    "grunt-contrib-uglify": "^0.9.1",
     "grunt-githooks": "^0.3.1",
     "grunt-jscs": "^2.0.0",
     "selenium-webdriver": "^2.48.0",
